* added a cast to remove a warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18991 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/geometry/TextSceneNode.cxx
@@ -446,7 +446,7 @@ TextSceneNode::TextRenderNode::TextRenderNode(TextSceneNode* _sceneNode,
 , text(*_text)
 , xformList(INVALID_GL_LIST_ID)
 , fontID(-1)
-, fontSize(text.fontSize)
+, fontSize((int)text.fontSize)
 , fixedWidth(0.0f)
 , noRadar(text.bzMaterial->getNoRadar())
 , noShadow(text.bzMaterial->getNoShadow())
